fix: clear cart

+++ src/order/v2/cart/v2/cart.service.js
@@ -80,7 +80,10 @@ class CartService {
         try {
             const cart = await Cart.findOne({userId:data.userId,_id:data.id})
             await Cart.deleteMany({userId:data.userId,_id:data.id})
-            return  await CartItem.deleteMany({cart:cart._id});
+            if(cart){
+                await CartItem.deleteMany({cart:cart._id});
+            }
+            return  {}
         }
         catch (err) {
             throw err;
